Reset Project
My original idea for the project was to create a pdf to audiobook application, however the several available libraries I've tested were unsuitable for the scope of this project due to root level setup requirements and bloated dependencies. (some of the trained data sets were reaching to GBs)
So for the sake of my sanity, I've decided to keep it simple and make a sudoku game.